SSL context; local WSDL to new with HTTPS URL
A custom X509TrustManager is created with Java's CAs + GoDaddy.  The manager
backs a new SSLContext that is used when the WebAPI stub is configured.

The 'cacerts' file contains the custom keystore.  The default password of
'changeit' is maintained.

A classpath WSDL is now provided as the bootstrap WSDL location.  The
endpoint property configured by the user is then set in the request context
of the WebAPI stub (BindingProvider) in order to properly route client
requests.

Java always finds a way to make things complicated.